Move Symfony integration
